+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Show ball-and-stick visual for amino acids linking carbohydrate to protein
+
 ## [2.1.1] - 2023-07-26
 
 - Fixed bug in checking expected files (do not check if empty when missing)